refac: move result screen files to own folder
