+++ src/components/nft-dashboard/Balance/components/SendForm/components/ResultScreen/ResultScreen.styles.ts
