+++ src/components/nft-dashboard/Balance/components/SendForm/components/ResultScreen/ResultScreen.tsx
